refactor(medications): split titrations-view into focused sub-components
titrations-view.tsx (1119 LOC) mixed plan list, plan card with phase
timeline, create/edit drawer, and Rx entry editor. Extracted to
src/components/medications/titrations/:

- maintenance-row.tsx: per-prescription maintenance summary
- titration-plan-card.tsx: collapsible plan card + nested PhaseEntryRow
- rx-entry-card.tsx: Rx entry editor + PrefillFromMaintenance +
  EditPhaseScheduleLoader (entry-related)
- titration-drawer.tsx: create/edit drawer shell
- use-titration-drawer-form.ts: form state hook (title, dates, notes,
  warnings, entries) with helpers and a prefillFromPlan setup
- types.ts: shared RxEntry interface and day labels

titrations-view.tsx is now 123 LOC and orchestrates these pieces.
Auth gating via useAuthGate and the AI titration-warnings call routed
through useAiFetch are preserved unchanged.

pnpm typecheck / test / lint / build all green.
