+++ src/components/medications/titrations/types.ts
@@ -1,0 +1,6 @@
+export interface RxEntry {
+  prescriptionId: string;
+  schedules: { time: string; daysOfWeek: number[]; dosage: string }[];
+}
+
+export const DAY_LABELS_LONG = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
